Add files via upload
Upload the ModelView files, connection to the model and to the view. Get element needed for binding from the model and give them to the view. Update model when the view is changed
Upload of the view files, with the different windows needed for our application. 
Upload of images used in our view